Added cmdline for kmeans clustering

+++ src/Dialogs/AdvancedPretreatmentDialog.cpp
@@ -109,7 +109,7 @@ void AdvancedPretreatmentDialog::OK()
   /* Now MX cpontain the pretreated matrix... */
   if(ui.addNoModVars->isChecked()){
     mx->MatrixResize(projects->value(pid)->getMatrix(mxid)->getObjName().size(), projects->value(pid)->getMatrix(mxid)->getVarName().size()-1);
-    for(size_t j = 1; j < projects->value(pid)->getMatrix(mxid)->getVarName().size(); j++){
+    for(int j = 1; j < projects->value(pid)->getMatrix(mxid)->getVarName().size(); j++){
       int indx = varnamestomod.indexOf(projects->value(pid)->getMatrix(mxid)->getVarName()[j]);
       mx->getVarName().append(projects->value(pid)->getMatrix(mxid)->getVarName()[j]);
       if(indx > -1){
